add auth middleware, reorganize code
