+++ controllers/main.js
@@ -15,25 +15,12 @@ const login = async (req, res)=> {
 }
 
 const dashboard = async (req, res) => {
-    const authHeader = req.headers.authorization;
-    if (!authHeader || !authHeader.startsWith('Bearer')) {
-        throw new CustomAPIError('No token provided', 401);
-    }
 
-    const token = authHeader.split(' ')[1]
-    console.log('Received token:', token);
-    try {
-        const decoded = jwt.verify(token, process.env.JWT_SECRET)
-        console.log('Decoded token:', decoded);
-        const luckyNumber = Math.floor(Math.random() * 100)
-        res.status(200).json({
-            msg: `Hello, ${decoded.username}`,
-            secret: `Here is your authorized data, you lucky number is ${luckyNumber}`
-        })
-    } catch (error) {
-        console.error('JWT Error:', error);
-        throw new CustomAPIError('Not authorized to access this route', 401)
-    }
+    const luckyNumber = Math.floor(Math.random() * 100)
+    res.status(200).json({
+        msg: `Hello, ${req.user.username}`,
+        secret: `Here is your authorized data, you lucky number is ${luckyNumber}`
+    })
 }
 
 module.exports = {
